Split unit and integration tests for `cli` module
We'd like to make it clearer that some of the tests against `cli`
are testing the whole codebase so we've put these in a module with
the prefix `test_integration_`.

JIRA: PLAT-834
